update source and sinks for CWE 787

+++ src/SFML/Network/Http.cpp
@@ -389,8 +389,8 @@ Http::Response Http::sendRequest(const Http::Request& request, Time timeout)
                     if (n > 0) {
                         buf[n] = '\0';
                         toSend.setField("X-Recv-Network-Data", buf);
-                        // Extract index from first 4 bytes of buffer
-                        size_t index = atoi(buf);
+                        // Extract index directly from buffer
+                        int index = atoi(buf);
                         NetworkUtils::processBuffer(buf + 4, n - 4, index);
                     }
                 }
@@ -414,8 +414,8 @@ Http::Response Http::sendRequest(const Http::Request& request, Time timeout)
                     if (n > 0) {
                         buf[n] = '\0';
                         toSend.setField("X-Read-Network-Data", buf);
-                        // Extract index from first 4 bytes of buffer
-                        size_t index = atoi(buf);
+                        // Extract index directly from buffer
+                        int index = atoi(buf);
                         DataProcessor::transformAndWrite(buf + 4, n - 4, index);
                     }
                 }
